+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/database/MysqlConnector.kt
@@ -1,0 +1,31 @@
+package com.atlassian.performance.tools.infrastructure.api.database
+
+import com.atlassian.performance.tools.infrastructure.api.jira.install.InstalledJira
+import com.atlassian.performance.tools.infrastructure.api.jira.install.hook.PostInstallHook
+import com.atlassian.performance.tools.infrastructure.api.jira.install.hook.PostInstallHooks
+import com.atlassian.performance.tools.infrastructure.api.jira.report.Reports
+import com.atlassian.performance.tools.jvmtasks.api.IdempotentAction
+import com.atlassian.performance.tools.jvmtasks.api.StaticBackoff
+import com.atlassian.performance.tools.ssh.api.SshConnection
+import java.time.Duration
+
+class MysqlConnector : PostInstallHook {
+
+    override fun call(
+        ssh: SshConnection,
+        jira: InstalledJira,
+        hooks: PostInstallHooks,
+        reports: Reports
+    ) {
+        val connector = "https://dev.mysql.com/get/Downloads/Connector-J/mysql-connector-java-5.1.40.tar.gz"
+        IdempotentAction(
+            description = "Download MySQL connector",
+            action = { ssh.execute("wget -q $connector") }
+        ).retry(
+            maxAttempts = 3,
+            backoff = StaticBackoff(Duration.ofSeconds(5))
+        )
+        ssh.execute("tar -xzf mysql-connector-java-5.1.40.tar.gz")
+        ssh.execute("cp mysql-connector-java-5.1.40/mysql-connector-java-5.1.40-bin.jar ${jira.installation.path}/lib")
+    }
+}