+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/network/Networked.kt
@@ -1,0 +1,9 @@
+package com.atlassian.performance.tools.infrastructure.api.network
+
+interface Networked {
+
+    /**
+     * @return CIDR
+     */
+    fun subnet(): String
+}